WIP, ner support

+++ main.py
@@ -4,10 +4,10 @@ import glob
 from modules.pdf_to_image import pdf2img
 from modules.make_dir import make_temp_dir, make_entity_dir
 from modules.image_to_txt import image_to_text
-from modules.entity_extractor import extract_resume_info
+from modules.entity_extractor import extract_resume_info_regex, extract_resume_info_NER
 
 
-def parse_all(docDir, txtDir, extractor):
+def parse_all(docDir, txtDir, preprocessing=False, extractor="regex"):
     for doc in os.listdir(docDir):
         try:
             fileExtension = doc.split(".")[-1]
@@ -17,7 +17,18 @@ def parse_all(docDir, txtDir, extractor):
                 temp_img_path = make_temp_dir(docDir)
                 pdfFilename = docDir + "/" + doc
                 pdf2img(pdfFilename, temp_img_path)
-                image_to_text(temp_img_path, txtDir, fileName)
+                if preprocessing:
+                    print("Preprocessing not implemented yet.")
+
+                if extractor == "NER":
+                    try:
+                        entDir = make_entity_dir(txtDir)
+                        extract_resume_info_NER(temp_img_path, entDir, fileName)
+                    except:
+                        print("Error finding entities from : " + str(txtDir))
+
+                else:
+                    image_to_text(temp_img_path, txtDir, fileName)
 
             else:
                 print("Only pdf files are supported.")
@@ -25,13 +36,14 @@ def parse_all(docDir, txtDir, extractor):
             print("Error in file: " + str(doc))
 
         finally:
-            for f in glob.glob(temp_img_path + "/*"):
-                os.remove(f)
+            if extractor != "NER":
+                for f in glob.glob(temp_img_path + "/*"):   
+                    os.remove(f)
 
     if extractor == "regex":
         try:
             entDir = make_entity_dir(txtDir)
-            extract_resume_info(txtDir, entDir)
+            extract_resume_info_regex(txtDir, entDir)
         except:
             print("Error finding entities from : " + str(txtDir))
 
@@ -40,6 +52,7 @@ if __name__ == "__main__":
     docDir = f"/home/punitsureka/Desktop/talview/OCR/tesseract/data/docs_to_parse"
     txtDir = f"/home/punitsureka/Desktop/talview/OCR/tesseract/data/docs_parsed"
 
-    extractor = "regex"  # or 'NER'
+    preprocessing = False  # or True
+    extractor = "NER"  # or 'NER'
 
-    parse_all(docDir, txtDir, extractor)
+    parse_all(docDir, txtDir, preprocessing, extractor)
